--- group_test.go
@@ -1,449 +1,0 @@
-package quick
-
-import (
-	"encoding/json"
-	"strings"
-	"testing"
-
-	"github.com/jeffotoni/quick/internal/concat"
-)
-
-// cover     ->  go test -v -count=1 -cover -failfast -run ^TestQuick_GroupGet$
-// coverHTML ->  go test -v -count=1 -failfast -cover -coverprofile=coverage.out -run ^TestQuick_GroupGet$; go tool cover -html=coverage.out
-func TestQuick_GroupGet(t *testing.T) {
-	type args struct {
-		route       string
-		wantCode    int
-		wantOut     string
-		isWantedErr bool
-		reqHeaders  map[string]string
-	}
-
-	type myType struct {
-		Name string `json:"name"`
-		Age  int    `json:"age"`
-	}
-
-	mt := myType{}
-	mt.Name = "jeff"
-	mt.Age = 35
-
-	testSuccessMockHandler := func(c *Ctx) error {
-		c.Set("Content-Type", "application/json")
-		return c.JSON(mt)
-	}
-
-	r := New()
-	g1 := r.Group("/v1/my")
-	g1.Get("/test", testSuccessMockHandler)
-	r.Get("/tester/:p1", testSuccessMockHandler)
-	r.Get("/", testSuccessMockHandler)
-	r.Get(`/my/reg/{(\S+)}/route`, testSuccessMockHandler)
-
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			name: "success",
-			args: args{
-				route:       "/v1/my/test?some=val1",
-				wantOut:     `{"name":"jeff","age":35}`,
-				wantCode:    200,
-				isWantedErr: false,
-			},
-		},
-		{
-			name: "success_with_params",
-			args: args{
-				route:       "/tester/val1",
-				wantOut:     `{"name":"jeff","age":35}`,
-				wantCode:    200,
-				isWantedErr: false,
-			},
-		},
-		{
-			name: "success_with_nothing",
-			args: args{
-				route:       "/",
-				wantOut:     `{"name":"jeff","age":35}`,
-				wantCode:    200,
-				isWantedErr: false,
-			},
-		},
-		{
-			name: "success_with_regex",
-			args: args{
-				route:       "/my/reg/88/route",
-				wantOut:     `{"name":"jeff","age":35}`,
-				wantCode:    200,
-				isWantedErr: false,
-			},
-		},
-		{
-			name: "error_not_exists_route",
-			args: args{
-				route:       "/tester/val1/route",
-				wantOut:     `404 page not found`,
-				wantCode:    404,
-				isWantedErr: true,
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-
-			data, err := r.QuickTest("GET", tt.args.route, tt.args.reqHeaders)
-			if (!tt.args.isWantedErr) && err != nil {
-				t.Errorf("error: %v", err)
-				return
-			}
-
-			s := strings.TrimSpace(data.BodyStr())
-			if s != tt.args.wantOut {
-				t.Errorf("was suppose to return %s and %s come", tt.args.wantOut, data.BodyStr())
-				return
-			}
-
-			if tt.args.wantCode != data.StatusCode() {
-				t.Errorf("was suppose to return %d and %d come", tt.args.wantCode, data.StatusCode())
-				return
-			}
-
-			t.Logf("outputBody -> %v", data.BodyStr())
-		})
-	}
-}
-
-// cover     ->  go test -v -count=1 -cover -failfast -run ^TestQuick_GroupPost$
-// coverHTML ->  go test -v -count=1 -failfast -cover -coverprofile=coverage.out -run ^TestQuick_GroupPost$; go tool cover -html=coverage.out
-func TestQuick_GroupPost(t *testing.T) {
-	type args struct {
-		route       string
-		wantCode    int
-		wantOut     string
-		isWantedErr bool
-		reqBody     []byte
-		reqHeaders  map[string]string
-	}
-
-	type myType struct {
-		Name string `json:"name"`
-		Age  int    `json:"age"`
-	}
-
-	testSuccessMockHandler := func(c *Ctx) error {
-		c.Set("Content-Type", "application/json")
-		b := c.Body()
-		resp := concat.String(`"data":`, string(b))
-		c.Status(200)
-		return c.SendString(resp)
-	}
-
-	testSuccessMockHandlerString := func(c *Ctx) error {
-		c.Set("Content-Type", "application/json")
-		mt := new(myType)
-		if err := c.BodyParser(mt); err != nil {
-			t.Errorf("error: %v", err)
-		}
-		b, _ := json.Marshal(mt)
-		resp := concat.String(`"data":`, string(b))
-		c.Status(200)
-		return c.String(resp)
-	}
-
-	testSuccessMockHandlerBind := func(c *Ctx) error {
-		c.Set("Content-Type", "application/json")
-		mt := new(myType)
-		if err := c.Bind(&mt); err != nil {
-			t.Errorf("error: %v", err)
-		}
-		b, _ := json.Marshal(mt)
-		resp := concat.String(`"data":`, string(b))
-		c.Status(200)
-		return c.String(resp)
-	}
-
-	r := New()
-	g1 := r.Group("/my/group")
-	g1.Post("/test", testSuccessMockHandler)
-	g1.Post("/tester/:p1", testSuccessMockHandler)
-	g2 := r.Group("/my")
-	g2.Post("/", testSuccessMockHandlerString)
-	r.Post("/bind", testSuccessMockHandlerBind)
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			name: "success",
-			args: args{
-				route:       "/my/group/test",
-				wantCode:    200,
-				wantOut:     `"data":{"name":"jeff", "age":35}`,
-				isWantedErr: false,
-				reqBody:     []byte(`{"name":"jeff", "age":35}`),
-			},
-		},
-		{
-			name: "success_param",
-			args: args{
-				route:       "/my/group/tester/:p1",
-				wantCode:    200,
-				wantOut:     `"data":{"name":"jeff", "age":35}`,
-				isWantedErr: false,
-				reqBody:     []byte(`{"name":"jeff", "age":35}`),
-			},
-		},
-		{
-			name: "success_without_param",
-			args: args{
-				route:       "/my/",
-				wantCode:    200,
-				wantOut:     `"data":{"name":"jeff","age":35}`,
-				isWantedErr: false,
-				reqBody:     []byte(`{"name":"jeff","age":35}`),
-				reqHeaders:  map[string]string{"Content-Type": "application/json"},
-			},
-		},
-		{
-			name: "success_bind",
-			args: args{
-				route:       "/bind",
-				wantCode:    200,
-				wantOut:     `"data":{"name":"jeff","age":35}`,
-				isWantedErr: false,
-				reqBody:     []byte(`{"name":"jeff","age":35}`),
-				reqHeaders:  map[string]string{"Content-Type": "application/json"},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-
-			data, err := r.QuickTest("POST", tt.args.route, tt.args.reqHeaders, tt.args.reqBody)
-			if (!tt.args.isWantedErr) && err != nil {
-				t.Errorf("error: %v", err)
-				return
-			}
-
-			s := strings.TrimSpace(data.BodyStr())
-			if s != tt.args.wantOut {
-				t.Errorf("was suppose to return %s and %s come", tt.args.wantOut, data.BodyStr())
-				return
-			}
-
-			if tt.args.wantCode != data.StatusCode() {
-				t.Errorf("was suppose to return %d and %d come", tt.args.wantCode, data.StatusCode())
-				return
-			}
-
-			t.Logf("outputBody -> %v", data.BodyStr())
-		})
-	}
-}
-
-// cover     -> go test -v -count=1 -cover -failfast -run ^TestQuick_PutGroup$
-// coverHTML -> go test -v -count=1 -failfast -cover -coverprofile=coverage.out -run ^TestQuick_PutGroup$; go tool cover -html=coverage.out
-func TestQuick_PutGroup(t *testing.T) {
-	type args struct {
-		route       string
-		wantCode    int
-		wantOut     string
-		isWantedErr bool
-		reqBody     []byte
-		reqHeaders  map[string]string
-	}
-
-	testSuccessMockHandler := func(c *Ctx) error {
-		c.Set("Content-Type", "application/json")
-		b := c.Body()
-		resp := concat.String(`"data":`, string(b))
-		return c.Byte([]byte(resp))
-	}
-
-	r := New()
-	r.Put("/", testSuccessMockHandler)
-	g1 := r.Group("/put/group")
-	g1.Put("/test", testSuccessMockHandler)
-	g1.Put("/tester/:p1", testSuccessMockHandler)
-	r.Put("/jeff", testSuccessMockHandler)
-
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			name: "success",
-			args: args{
-				route:       "/put/group/test",
-				wantCode:    200,
-				wantOut:     `"data":{"name":"jeff", "age":35}`,
-				isWantedErr: false,
-				reqBody:     []byte(`{"name":"jeff", "age":35}`),
-			},
-		},
-		{
-			name: "success_param",
-			args: args{
-				route:       "/put/group/tester/:p1",
-				wantCode:    200,
-				wantOut:     `"data":{"name":"jeff", "age":35}`,
-				isWantedErr: false,
-				reqBody:     []byte(`{"name":"jeff", "age":35}`),
-			},
-		},
-		{
-			name: "success_without_param",
-			args: args{
-				route:       "/",
-				wantCode:    200,
-				wantOut:     `"data":{"name":"jeff", "age":35}`,
-				isWantedErr: false,
-				reqBody:     []byte(`{"name":"jeff", "age":35}`),
-				reqHeaders:  map[string]string{"Content-Type": "application/json"},
-			},
-		},
-		{
-			name: "success_without_param",
-			args: args{
-				route:       "/jeff",
-				wantCode:    200,
-				wantOut:     `"data":{"name":"jeff", "age":35}`,
-				isWantedErr: false,
-				reqBody:     []byte(`{"name":"jeff", "age":35}`),
-				reqHeaders:  map[string]string{"Content-Type": "application/json"},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			data, err := r.QuickTest("PUT", tt.args.route, tt.args.reqHeaders, tt.args.reqBody)
-			if (!tt.args.isWantedErr) && err != nil {
-				t.Errorf("error: %v", err)
-				return
-			}
-
-			s := strings.TrimSpace(data.BodyStr())
-			if s != tt.args.wantOut {
-				t.Errorf("route %s -> was suppose to return %s and %s come", tt.args.route, tt.args.wantOut, data.BodyStr())
-				return
-			}
-
-			if tt.args.wantCode != data.StatusCode() {
-				t.Errorf("was suppose to return %d and %d come", tt.args.wantCode, data.StatusCode())
-				return
-			}
-
-			t.Logf("\nOutputBodyString -> %v", data.BodyStr())
-			t.Logf("\nStatusCode -> %d", data.StatusCode())
-			t.Logf("\nOutputBody -> %v", string(data.Body())) // I have converted in this example to string but comes []byte as default
-			t.Logf("\nResponse -> %v", data.Response())
-		})
-	}
-}
-
-// cover     -> go test -v -count=1 -cover -failfast -run ^TestQuick_DeleteGroup$
-// coverHTML -> go test -v -count=1 -failfast -cover -coverprofile=coverage.out -run ^TestQuick_DeleteGroup$; go tool cover -html=coverage.out
-func TestQuick_DeleteGroup(t *testing.T) {
-	type args struct {
-		route       string
-		wantOut     string
-		wantCode    int
-		isWantedErr bool
-		reqHeaders  map[string]string
-	}
-
-	testSuccessMockHandler := func(c *Ctx) error {
-		c.Set("Content-Type", "application/json")
-		return c.Status(204).SendString("")
-	}
-
-	r := New()
-	r.Delete("/", testSuccessMockHandler)
-	g1 := r.Group("/del/group")
-	g1.Delete("/user", testSuccessMockHandler)
-	g1.Delete("/tester/:p1", testSuccessMockHandler)
-	r.Delete("/jeff", testSuccessMockHandler)
-
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			name: "success",
-			args: args{
-				route:       "/",
-				wantOut:     ``,
-				wantCode:    204,
-				isWantedErr: false,
-			},
-		},
-		{
-			name: "success",
-			args: args{
-				route:       "/del/group/user",
-				wantOut:     ``,
-				wantCode:    204,
-				isWantedErr: false,
-			},
-		},
-		{
-			name: "success_param",
-			args: args{
-				route:       "/del/group/tester/:p1",
-				wantOut:     ``,
-				wantCode:    204,
-				isWantedErr: false,
-			},
-		},
-		{
-			name: "success_without_param",
-			args: args{
-				route:       "/jeff",
-				wantOut:     ``,
-				wantCode:    204,
-				isWantedErr: false,
-				reqHeaders:  map[string]string{"Content-Type": "application/json"},
-			},
-		},
-		{
-			name: "success_without_param",
-			args: args{
-				route:       "/nao/existe/esta/rota",
-				wantOut:     "404 page not found",
-				wantCode:    404,
-				isWantedErr: false,
-				reqHeaders:  map[string]string{"Content-Type": "application/json"},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			data, err := r.QuickTest("DELETE", tt.args.route, tt.args.reqHeaders)
-			if (!tt.args.isWantedErr) && err != nil {
-				t.Errorf("error: %v", err)
-				return
-			}
-
-			s := strings.TrimSpace(data.BodyStr())
-			if s != tt.args.wantOut {
-				t.Errorf("was suppose to return %s and %s come", tt.args.wantOut, data.BodyStr())
-				return
-			}
-
-			if tt.args.wantCode != data.StatusCode() {
-				t.Errorf("was suppose to return %d and %d come", tt.args.wantCode, data.StatusCode())
-				return
-			}
-
-			t.Logf("outputBody -> %v", data.BodyStr())
-
-			// t.Logf("\nOutputBodyString -> %v", data.BodyStr())
-			// t.Logf("\nStatusCode -> %d", data.StatusCode())
-			// t.Logf("\nOutputBody -> %v", string(data.Body())) // I have converted in this example to string but comes []byte as default
-			// t.Logf("\nResponse -> %v", data.Response())
-		})
-	}
-}
